fix(BRE): small refactorings
