+++ core/services/business-rules-executor/index.js
@@ -1,7 +1,7 @@
 const { setLoggerDefaultMeta, logger: log } = require("@diva/common/logger");
 const generateUuid = require("@diva/common/generateUuid");
 const serviceName = require("./package.json").name;
-const eventsHandlerService = require("./services/EventsHandlerService");
+const eventsConsumerService = require("./services/EventsConsumerService");
 
 const serviceId = generateUuid("service");
 setLoggerDefaultMeta({ serviceId });
@@ -10,7 +10,7 @@ const NODE_ENV = process.env.NODE_ENV || "development";
 
 log.info(`✅ Booting ${serviceName} in ${NODE_ENV} mode`);
 
-eventsHandlerService
+eventsConsumerService
   .init()
   .then(() => log.info(`✅ All components booted successfully 🚀`))
   .catch((e) => {
